test: split test_valor_telegram into per-theme modules and absorb its siblings (Refs #2879)
