+++ tests/unit/valor_telegram/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the unified valor-telegram CLI tool, split by theme (#2879)."""
